refactor: material && page service
