+++ extensions/iceworks-lowcode-page-builder/ext-src/services/index.ts
@@ -1,4 +1,4 @@
 import * as material from './material';
-import * as page from './page/index';
+import * as page from './page';
 
 export default { material, page };
