+++ extensions/iceworks-lowcode-page-builder/ext-src/services/page.ts
@@ -1,0 +1,1 @@
+export * from '@iceworks/page-service';
