+++ extensions/iceworks-lowcode-page-builder/package.json
@@ -94,14 +94,13 @@
   "dependencies": {
     "@alifd/next": "^1.19.4",
     "@alifd/theme-design-pro": "^0.x",
-    "@iceworks/material": "^0.1.0",
+    "@iceworks/material-service": "0.1.0-beta-2",
+    "@iceworks/page-service": "0.1.0-beta-2",
+    "@iceworks/material-ui": "0.1.0-beta-2",
     "axios": "^0.19.2",
-    "ejs": "^3.1.3",
     "moment": "^2.24.0",
-    "prettier": "^2.0.5",
     "react": "^16.4.1",
     "react-dom": "^16.4.1",
-    "react-lazyload": "^2.6.7",
     "react-sortable-hoc": "^1.11.0"
   }
 }
