--- extensions/iceworks-lowcode-page-builder/src/components/Material/index.module.scss
@@ -1,3 +1,0 @@
-.componentMaterialWrap {
-  height: 100%;
-}
