+++ packages/material-service/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@iceworks/material",
-  "version": "0.1.0",
+  "name": "@iceworks/material-service",
+  "version": "0.1.0-beta-2",
   "files": [
     "lib"
   ],
@@ -10,5 +10,8 @@
     "axios": "^0.19.2",
     "lodash.kebabcase": "^4.1.1",
     "package-json": "^6.5.0"
+  },
+  "devDependencies": {
+    "@types/vscode": "^1.45.0"
   }
 }
