+++ packages/material-ui/.eslintignore
@@ -1,0 +1,6 @@
+# 忽略目录
+build/
+node_modules/
+**/*-min.js
+**/*.min.js
+coverage/
