+++ packages/material-ui/.eslintrc.js
@@ -1,0 +1,5 @@
+const { eslint, deepmerge } = require('@ice/spec');
+
+module.exports = deepmerge(eslint, {
+  rules: {},
+});
