+++ packages/material-ui/README.md
@@ -1,0 +1,11 @@
+# material demo
+
+material
+
+物料面板
+
+## API
+
+| 参数名 | 说明 | 必填 | 类型 | 默认值 | 备注 |
+| ------ | ---- | ---- | ---- | ------ | ---- |
+|        |      |      |      |        |      |
