+++ packages/material-ui/package.json
@@ -1,0 +1,57 @@
+{
+  "name": "@iceworks/material-ui",
+  "version": "0.1.0-beta-2",
+  "description": "物料面板",
+  "files": [
+    "demo/",
+    "es/",
+    "lib/",
+    "build/"
+  ],
+  "main": "lib/index.js",
+  "module": "es/index.js",
+  "stylePath": "style.js",
+  "scripts": {
+    "start": "build-scripts start",
+    "build": "build-scripts build",
+    "prepublishOnly": "npm run build",
+    "lint": "eslint --cache --ext .js,.jsx ./"
+  },
+  "keywords": [
+    "ice",
+    "react",
+    "component"
+  ],
+  "dependencies": {
+    "prop-types": "^15.5.8",
+    "react-lazyload": "^2.6.7",
+    "@iceworks/material-utils": "^0.1.0-beta-2"
+  },
+  "devDependencies": {
+    "@alib/build-scripts": "^0.1.3",
+    "build-plugin-component": "^0.2.0",
+    "build-plugin-fusion": "^0.1.0",
+    "build-plugin-moment-locales": "^0.1.0",
+    "react": "^16.3.0",
+    "react-dom": "^16.3.0",
+    "@ice/spec": "^0.1.1",
+    "eslint": "^6.0.1",
+    "@alifd/next": "1.x",
+    "@types/react": "^16.9.13",
+    "@types/react-dom": "^16.9.4"
+  },
+  "peerDependencies": {
+    "react": "^16.3.0",
+    "@alifd/next": "1.x"
+  },
+  "componentConfig": {
+    "name": "material",
+    "title": "material demo",
+    "category": "Others"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "license": "MIT",
+  "homepage": "https://unpkg.com/@iceworks/material-ui@0.1.0-beta-2/build/index.html"
+}
