+++ packages/material-ui/src/components/base.module.scss
@@ -1,3 +1,5 @@
+@import '~@alifd/next/lib/core/index.scss';
+
 .container {
   transition: transform 0.3s ease;
   display: block;
@@ -34,7 +36,7 @@
 .title {
   margin: 0;
   padding: 10px;
-  border-top: 1px solid $color-line1-2;
+  border-top: 1px solid #4D4F61;
   text-align: center;
 }
