+++ packages/material-ui/src/components/component.module.scss
@@ -1,11 +1,15 @@
+@import '~@alifd/next/lib/core/index.scss';
+
 .container {
+  position: relative;
   transition: transform 0.3s ease;
   display: block;
   cursor: pointer;
+  padding: 16px;
+  text-align: left;
   overflow: hidden;
-  position: relative;
-  height: 100%;
-  margin-bottom: 24px;
+  height: 100px;
+  margin-bottom: 14px;
   border: 1px solid $color-line1-1;
   &:hover {
     text-decoration: none;
@@ -16,26 +20,18 @@
   }
 }
 
-.screenshot {
-  position: relative;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 160px;
-  transition: all 0.3s ease;
-
-  img {
-    max-width: 100%;
-    max-height: 100%;
-    display: block;
-  }
+.title {
+  margin: 0 0 5px;
+  height: 30px;
+  line-height: 1.6;
+  font-weight: 500;
 }
 
-.title {
-  margin: 0;
-  padding: 10px;
-  border-top: 1px solid #4D4F61;
-  text-align: center;
+.desc {
+  font-size: 12px;
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
 }
 
 .actions {
@@ -56,8 +52,8 @@
     align-items: center;
     justify-content: center;
     width: 50%;
-    height: 45px;
-    line-height: 45px;
+    height: 35px;
+    line-height: 35px;
     font-size: 14px;
     font-weight: bold;
     background: $color-fill1-3;
