+++ packages/material-ui/src/components/scaffold.module.scss
@@ -1,3 +1,5 @@
+@import '~@alifd/next/lib/core/index.scss';
+
 .scaffold {
   position: relative;
   display: block;
