+++ packages/material-ui/src/components/type.module.scss
@@ -1,3 +1,5 @@
+@import '~@alifd/next/lib/core/index.scss';
+
 .spin {
   width: 100%;
   height: 100%;
