+++ packages/material-ui/src/components/view.module.scss
