+++ packages/material-ui/src/components/view.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { Tab } from '@alifd/next';
-import { IMaterialSource, IMaterialTypeDatum, IMaterialScaffold, IMaterialBlock, IMaterialComponent, IMaterialBase } from '@iceworks/material/lib/common';
+import { IMaterialSource, IMaterialTypeDatum, IMaterialScaffold, IMaterialBlock, IMaterialComponent, IMaterialBase } from '@iceworks/material-utils';
 import { MaterialType } from './type';
 import * as styles from './view.module.scss';
