+++ packages/material-ui/src/index.scss
@@ -1,0 +1,3 @@
+.iceworks-material {
+  height: 100%;
+}
