+++ packages/material-utils/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "@iceworks/material-utils",
+  "version": "0.1.0-beta-2",
+  "files": [
+    "lib"
+  ],
+  "main": "lib/index.js",
+  "typings": "lib/index.d.ts"
+}
