--- packages/material/src/index.ts
@@ -1,2 +1,0 @@
-export * from './common';
-export * from './node';
