+++ packages/page-service/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "@iceworks/page-service",
+  "version": "0.1.0-beta-2",
+  "files": [
+    "lib"
+  ],
+  "main": "lib/index.js",
+  "typings": "lib/index.d.ts",
+  "dependencies": {
+    "@iceworks/material-utils": "^0.1.0-beta-2",
+    "ejs": "^3.1.3",
+    "fs-extra": "^9.0.0",
+    "ice-npm-utils": "^1.4.1",
+    "prettier": "^2.0.5",
+    "uppercamelcase": "^3.0.0"
+  },
+  "devDependencies": {
+    "@types/vscode": "^1.45.0"
+  }
+}
