+++ tsconfig.json
@@ -6,6 +6,7 @@
     {"path": "packages/generate-project"},
     {"path": "packages/block-add"},
     {"path": "packages/iceworks-cli"},
-    {"path": "packages/material"}
+    {"path": "packages/material-utils"},
+    {"path": "packages/material-service"}
   ]
 }
